fix errors in cpp00

+++ CPP00/ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: aboukdid <aboukdid@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/04 15:43:07 by aboukdid          #+#    #+#             */
-/*   Updated: 2024/08/04 19:52:31 by aboukdid         ###   ########.fr       */
+/*   Updated: 2024/10/03 14:25:27 by aboukdid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int main(int argc, char **argv)
 		for (int i = 1; i < argc ; i++)
 		{
 			str = argv[i];
-			for (unsigned long j = 0; j < str.length(); j++)
+			for (size_t j = 0; j < str.length(); j++)
 				str[j] = std::toupper(str[j]);
 			std::cout << str;
 		}
